\name{amelia-package}
\alias{amelia-package}
\docType{package}
\title{Multiple Impuation of Incomplete Multivariate Data}
\description{Uses a bootstrap EM algorithm to impute missing values
  from a dataset and produce multiple output datasets for analysis.}
\details{
\tabular{ll}{
Package: \tab amelia\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2006-03-03\cr
License: \tab See Manual\cr
}
You can use the package in one of two ways:  either by invoking the 
\code{ameliagui()} command and running the program from a graphical 
interface or by loading in your data and then running the \code{amelia}
function on the data.  Both methods are documented in the Amelia manual.

If you use the GUI in Windows, makes sure that you run R under a Single
Window Interface (SDI) as it will try to grab focus from the GUI if
you don't.
}
\author{
James Honaker and Matthew Blackwell
}
\references{
King, Gary; James Honaker, Anne Joseph, and Kenneth Scheve. \"Analyzing 
Incomplete Political Science Data: An Alternative Algorithm for Multiple 
Imputation\", \emph{American Political Science Review}, Vol. 95, No. 1 
(March, 2001): Pp. 49-69.
}
\keyword{package}
