% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{amelia_colors}
\alias{amelia_colors}
\title{Function to extract amelia colors as hex codes}
\usage{
amelia_colors(...)
}
\arguments{
\item{...}{Character names of cb_colors}
}
\description{
Function to extract amelia colors as hex codes
}
