\name{amelia.gui}
\alias{amelia.gui}
\title{Interactive GUI for Amelia}
\description{Brings up the Amelia graphical interface, which allows users
  to load datasets, manage options and run Amelia from a traditional 
  windowed environment.
}
\usage{
amelia.gui()
}
\keyword{ utilities }% at least one, from doc/KEYWORDS
