% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ameliagui.r
\name{AmeliaView}
\alias{AmeliaView}
\title{Interactive GUI for Amelia}
\usage{
AmeliaView()
}
\description{
Brings up the AmeliaView graphical interface, which allows users to load datasets, 
manage options and run Amelia from a traditional windowed environment.
}
\details{
Requires the tcltk package.
}
