% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bundesland}
\alias{bundesland}
\title{Geospatial data of Bunderländer in Germany}
\format{An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 18320 rows and 6 columns.}
\usage{
bundesland
}
\description{
A tidy dataframe containing the latitude and longitudes of Bunderländer
in Germany for use with ggplot2::geom_polygon.
}
\details{
\describe{
  \item{long}{longitude}
  \item{lat}{latitutde}
  \item{id}{Numeric code used as governmental ID and prefix for regional codes}
  \item{gen}{Bundesland name}
  \item{hole}{Logical. Whether polygon is a hole or not}
  ...
}
}
\keyword{datasets}
