% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{fsu_dark_colors}
\alias{fsu_dark_colors}
\title{Function to extract FSU dark colors as hex codes}
\usage{
fsu_dark_colors(...)
}
\arguments{
\item{...}{Character names of uni jena dark colors}
}
\description{
Function to extract FSU dark colors as hex codes
}
