% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{fsu_light_colors}
\alias{fsu_light_colors}
\title{Function to extract FSU light colors as hex codes}
\usage{
fsu_light_colors(...)
}
\arguments{
\item{...}{Character names of uni jena light colors}
}
\description{
Function to extract FSU light colors as hex codes
}
