% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{gm_mean}
\alias{gm_mean}
\title{Geometric mean}
\usage{
gm_mean(x, na.rm = TRUE)
}
\arguments{
\item{x}{a numeric vector}

\item{na.rm}{logical. Whether missing values should be removed.}
}
\description{
This function calculates the geometric mean of the numeric vector `x`.
}
\keyword{geometric}
\keyword{mean}
