% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{landkreis}
\alias{landkreis}
\title{Geospatial data of Landkreise in Germany}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 60889 rows and 8 columns.
}
\usage{
landkreis
}
\description{
A tidy dataframe containing the latitude and longitudes of Landkreise
in Germany for use with ggplot2::geom_polygon.
}
\details{
\describe{
  \item{long}{longitude}
  \item{lat}{latitutde}
  \item{rs}{Landkreise code}
  \item{gen}{Landkreis name}
  \item{hole}{Logical. Whether polygon is a hole or not}
  ...
}
}
\keyword{datasets}
