% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{mi.combine}
\alias{mi.combine}
\title{Combine results from analyses on imputed data sets}
\usage{
mi.combine(x, conf.int = FALSE, conf.level = 0.95)
}
\arguments{
\item{x}{List of output from statistical models estimated on
different imputed data sets, as outputted by \code{with(a.out,
expr)} where \code{a.out} is the output of a call to \code{amelia}.}

\item{conf.int}{Logical indicating if confidence intervals should
be computed for each quantity of interest (default is \code{FALSE}).}

\item{conf.level}{The confidence level to use for the confidence
interval  if \code{conf.level = TRUE}. Defaults to 0.95, which
corresponds to a 95 percent confidence interval.}
}
\value{

}
\description{
Combine results from statistical models run on multiply imputed
data sets using the so-called Rubin rules.
}
\author{
Matt Blackwell
}
