% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{microverse_colors}
\alias{microverse_colors}
\title{Function to extract microverse colors as hex codes}
\usage{
microverse_colors(...)
}
\arguments{
\item{...}{Character names of cb_colors}
}
\description{
Function to extract microverse colors as hex codes
}
