% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{panel_border}
\alias{panel_border}
\title{Add/remove the panel border in a ggplot2 plot}
\usage{
panel_border(color = "grey50", size = 1, linetype = 1,
  remove = FALSE, colour)
}
\arguments{
\item{color, colour}{The color of the border.}

\item{size}{Size. Needs to be twice as large as desired outcome when panel clipping
is on (the default).}

\item{linetype}{Line type.}

\item{remove}{If \code{TRUE}, removes the current panel border.}
}
\description{
This function adds or removes the panel border in ggplot2.
Modified from Cowplot
}
