% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{scale_color_amelia}
\alias{scale_color_amelia}
\title{Function to provide color blind-friendly scales}
\usage{
scale_color_amelia(..., palette = "cb_grey")
}
\arguments{
\item{palette}{Name of color palette to use (currently, "cb_grey" ir "cb_black")}
}
\description{
Function to provide color blind-friendly scales
}
