% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{scale_color_fsu}
\alias{scale_color_fsu}
\title{Function to provide color fsu theme colors to ggplot}
\usage{
scale_color_fsu(..., palette = "fsu_dark")
}
\arguments{
\item{palette}{Name of color palette to use (currently, "fsu_dark" or "fsu_light")}
}
\description{
Function to provide color fsu theme colors to ggplot
}
