% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{scale_color_microverse}
\alias{scale_color_microverse}
\title{Function to provide microverse palette to scale_color functions}
\usage{
scale_color_microverse(...)
}
\description{
Function to provide microverse palette to scale_color functions
}
