% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{scale_fill_fsu}
\alias{scale_fill_fsu}
\title{Function to provide color blind-friendly fills}
\usage{
scale_fill_fsu(..., palette = "fsu_dark")
}
\arguments{
\item{palette}{Name of color palette to use (currently, either "cb_grey" or "cb_black")}
}
\description{
Function to provide color blind-friendly fills
}
