% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{scale_fill_microverse}
\alias{scale_fill_microverse}
\title{Function to provide microverse palette to scale_fill functions}
\usage{
scale_fill_microverse(...)
}
\description{
Function to provide microverse palette to scale_fill functions
}
