% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{se}
\alias{se}
\title{Standard error}
\usage{
se(x, na.rm = TRUE)
}
\arguments{
\item{x}{a numeric vector}

\item{na.rm}{logical. Whether missing values should be removed.}
}
\description{
This function calculates the standard error of all the values in `x`.
}
\examples{



}
\keyword{error}
\keyword{standard}
