% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.amelia.R
\name{summary.amelia}
\alias{summary.amelia}
\title{Summary of an Amelia object}
\usage{
\method{summary}{amelia}(object, ...)
}
\arguments{
\item{object}{an object of class \code{amelia}. Typically, an output
from the function \code{amelia}.}

\item{...}{further arguments.}
}
\description{
Returns summary information from the Amelia run along with
missingles information.
}
\seealso{
\code{\link{amelia}}, \code{\link{plot.amelia}}
}
