% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.mi.R
\name{summary.mi}
\alias{summary.mi}
\title{Summary of an mi object}
\usage{
\method{summary}{mi}(object, ...)
}
\arguments{
\item{object}{an object of class \code{mi}. Typically, an output
from the function \code{amelia}.}

\item{...}{further arguments.}
}
\description{
Returns summary information about the list of multiply imputed data
sets
}
\seealso{
\code{\link{amelia}}
}
