% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_halfopen}
\alias{theme_halfopen}
\title{Theme from cowplot}
\usage{
theme_halfopen(
  font_size = 14,
  font_family = "",
  line_size = 0.5,
  rel_small = 12/14,
  rel_tiny = 11/14,
  rel_large = 16/14
)
}
\description{
https://github.com/wilkelab/cowplot
}
