% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_minimal_grid}
\alias{theme_minimal_grid}
\alias{theme_minimal_vgrid}
\alias{theme_minimal_hgrid}
\title{Minimalistic themes with grids from Cowplot
https://github.com/wilkelab/cowplot}
\usage{
theme_minimal_grid(font_size = 14, font_family = "", line_size = 0.5,
  rel_small = 12/14, rel_tiny = 11/14, rel_large = 16/14,
  color = "grey85", colour)

theme_minimal_vgrid(font_size = 14, font_family = "",
  line_size = 0.5, rel_small = 12/14, rel_tiny = 11/14,
  rel_large = 16/14, color = "grey85", colour)

theme_minimal_hgrid(font_size = 14, font_family = "",
  line_size = 0.5, rel_small = 12/14, rel_tiny = 11/14,
  rel_large = 16/14, color = "grey85", colour)
}
\description{
Three minimalistic themes that provide either a full grid,
a horizontal grid, or a vertical grid.
}
\details{
`theme_minimal_grid()` provides a minimal grid theme. `theme_minimal_hgrid()` strips down
this theme even further and draws only horizontal lines, and `theme_minimal_vgrid()`
does the same for vertical lines.
}
