\name{africa}
\docType{data}
\alias{africa}

\title{Economic and Political Indictors in 6 African States}

\description{Data on a few economic and political variables in six
  African States from 1972-1991. The variables are year, country name,
  Gross Domestic Product per capita, inflation, trade as a percentage of
  GDP, a measure of civil liberties and total population. The data is
  from the Africa Research Program. A few cells are missing.}

\usage{africa}
\format{A data frame with 7 variables and 120 observations.}
\source{Africa Research Program \url{https://scholar.harvard.edu/rbates/data}}
\keyword{datasets}
