% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.amelia.R
\name{plot.amelia}
\alias{plot.amelia}
\title{Summary plots for Amelia objects}
\usage{
\method{plot}{amelia}(x, which.vars, compare = TRUE, overimpute = FALSE,
  ask = TRUE, ...)
}
\arguments{
\item{x}{an object of class "amelia"; typically output from the
function \code{amelia}.}

\item{which.vars}{a vector indicating the variables to plot. The
default is to plot all of the numeric variables that were actually
imputed.}

\item{compare}{plot the density comparisons for each variable (True or False)}

\item{overimpute}{plot the overimputation for each variable (True or False)}

\item{ask}{prompt user before changing pages of a plot (True or False)}

\item{...}{further graphical arguments.}
}
\description{
Plots diagnostic plots for the output from the
\code{amelia} function.
}
